mv coding plans to configurations
