+++ src/lib/pipeline_configuration.py
@@ -4,209 +4,18 @@ from datetime import datetime
 from urllib.parse import urlparse
 
 import pytz
-from core_data_modules.cleaners import Codes, swahili, somali
 from core_data_modules.data_models import validators
-from core_data_modules.traced_data.util.fold_traced_data import FoldStrategies
 from dateutil.parser import isoparse
 
-from src.lib import CodeSchemes, code_imputation_functions
-
-
-class CodingModes(object):
-    SINGLE = "SINGLE"
-    MULTIPLE = "MULTIPLE"
-
-
-class CodingConfiguration(object):
-    def __init__(self, coding_mode, code_scheme, coded_field, fold_strategy, analysis_file_key=None, cleaner=None):
-        assert coding_mode in {CodingModes.SINGLE, CodingModes.MULTIPLE}
-
-        self.coding_mode = coding_mode
-        self.code_scheme = code_scheme
-        self.coded_field = coded_field
-        self.analysis_file_key = analysis_file_key
-        self.fold_strategy = fold_strategy
-        self.cleaner = cleaner
-
-
-class CodingPlan(object):
-    def __init__(self, raw_field, coding_configurations, raw_field_fold_strategy, coda_filename=None, ws_code=None,
-                 time_field=None, run_id_field=None, icr_filename=None, id_field=None, code_imputation_function=None):
-        self.raw_field = raw_field
-        self.time_field = time_field
-        self.run_id_field = run_id_field
-        self.coda_filename = coda_filename
-        self.icr_filename = icr_filename
-        self.coding_configurations = coding_configurations
-        self.code_imputation_function = code_imputation_function
-        self.ws_code = ws_code
-        self.raw_field_fold_strategy = raw_field_fold_strategy
-
-        if id_field is None:
-            id_field = "{}_id".format(self.raw_field)
-        self.id_field = id_field
+from configuration import coding_plans
 
 
 class PipelineConfiguration(object):
-
-    RQA_CODING_PLANS =  [
-        CodingPlan(raw_field="rqa_s01e01_raw",
-                   time_field="sent_on",
-                   run_id_field="rqa_s01e01_run_id",
-                   coda_filename="COVID19_SOM_s01e01.json",
-                   icr_filename="s01e01.csv",
-                   coding_configurations=[
-                       CodingConfiguration(
-                           coding_mode=CodingModes.MULTIPLE,
-                           code_scheme=CodeSchemes.S01E01,
-                           coded_field="rqa_s01e01_coded",
-                           analysis_file_key="rqa_s01e01_",
-                           fold_strategy=lambda x, y: FoldStrategies.list_of_labels(CodeSchemes.S01E01, x, y)
-                       )
-                   ],
-                   ws_code=CodeSchemes.WS_CORRECT_DATASET.get_code_with_match_value("covid19 som s01e01"),
-                   raw_field_fold_strategy=FoldStrategies.concatenate),
-
-    ]
-
-    @staticmethod
-    def clean_age_with_range_filter(text):
-        """
-        Cleans age from the given `text`, setting to NC if the cleaned age is not in the range 10 <= age < 100.
-        """
-        age = somali.DemographicCleaner.clean_age(text)
-        if type(age) == int and 10 <= age < 100:
-            return str(age)
-            # TODO: Once the cleaners are updated to not return Codes.NOT_CODED, this should be updated to still return
-            #       NC in the case where age is an int but is out of range
-        else:
-            return Codes.NOT_CODED
-
-    @staticmethod
-    def clean_district_if_no_mogadishu_sub_district(text):
-        mogadishu_sub_district = somali.DemographicCleaner.clean_mogadishu_sub_district(text)
-        if mogadishu_sub_district == Codes.NOT_CODED:
-            return somali.DemographicCleaner.clean_somalia_district(text)
-        else:
-            return Codes.NOT_CODED
-
-    SURVEY_CODING_PLANS = [
-        CodingPlan(raw_field="gender_raw",
-                   time_field="gender_time",
-                   coda_filename="IMAQAL_gender.json",
-                   coding_configurations=[
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.GENDER,
-                           cleaner=somali.DemographicCleaner.clean_gender,
-                           coded_field="gender_coded",
-                           analysis_file_key="gender",
-                           fold_strategy=FoldStrategies.assert_label_ids_equal
-                       )
-                   ],
-                   ws_code=CodeSchemes.WS_CORRECT_DATASET.get_code_with_match_value("gender"),
-                   raw_field_fold_strategy=FoldStrategies.assert_equal),
-
-        CodingPlan(raw_field="age_raw",
-                   time_field="age_time",
-                   coda_filename="IMAQAL_age.json",
-                   coding_configurations=[
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.AGE,
-                           cleaner=lambda text: PipelineConfiguration.clean_age_with_range_filter(text),
-                           coded_field="age_coded",
-                           analysis_file_key="age",
-                           fold_strategy=FoldStrategies.assert_label_ids_equal
-                       ),
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.AGE_CATEGORY,
-                           coded_field="age_category_coded",
-                           analysis_file_key="age_category",
-                           fold_strategy=FoldStrategies.assert_label_ids_equal
-                       )
-                   ],
-                   code_imputation_function=code_imputation_functions.impute_age_category,
-                   ws_code=CodeSchemes.WS_CORRECT_DATASET.get_code_with_match_value("age"),
-                   raw_field_fold_strategy=FoldStrategies.assert_equal),
-
-        CodingPlan(raw_field="recently_displaced_raw",
-                   time_field="recently_displaced_time",
-                   coda_filename="IMAQAL_recently_displaced.json",
-                   coding_configurations=[
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.RECENTLY_DISPLACED,
-                           cleaner=somali.DemographicCleaner.clean_yes_no,
-                           coded_field="recently_displaced_coded",
-                           analysis_file_key="recently_displaced",
-                           fold_strategy=FoldStrategies.assert_label_ids_equal
-                       )
-                   ],
-                   ws_code=CodeSchemes.WS_CORRECT_DATASET.get_code_with_match_value("recently displaced"),
-                   raw_field_fold_strategy=FoldStrategies.assert_equal),
-
-        CodingPlan(raw_field="household_language_raw",
-                   time_field="household_language_time",
-                   coda_filename="IMAQAL_household_language.json",
-                   coding_configurations=[
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.HOUSEHOLD_LANGUAGE,
-                           coded_field="household_language_coded",
-                           analysis_file_key="household_language",
-                           fold_strategy=FoldStrategies.assert_label_ids_equal
-                       )
-                   ],
-                   ws_code=CodeSchemes.WS_CORRECT_DATASET.get_code_with_match_value("hh language"),
-                   raw_field_fold_strategy=FoldStrategies.assert_equal),
-
-        CodingPlan(raw_field="location_raw",
-                   time_field="location_time",
-                   coda_filename="IMAQAL_location.json",
-                   coding_configurations=[
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.MOGADISHU_SUB_DISTRICT,
-                           fold_strategy=FoldStrategies.assert_label_ids_equal,
-                           cleaner=somali.DemographicCleaner.clean_mogadishu_sub_district,
-                           coded_field="mogadishu_sub_district_coded"
-                       ),
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.SOMALIA_DISTRICT,
-                           cleaner=lambda text: PipelineConfiguration.clean_district_if_no_mogadishu_sub_district(text),
-                           fold_strategy=FoldStrategies.assert_label_ids_equal,
-                           coded_field="district_coded",
-                           analysis_file_key="district"
-                       ),
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.SOMALIA_REGION,
-                           fold_strategy=FoldStrategies.assert_label_ids_equal,
-                           coded_field="region_coded",
-                           analysis_file_key="region",
-                       ),
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.SOMALIA_STATE,
-                           fold_strategy=FoldStrategies.assert_label_ids_equal,
-                           coded_field="state_coded",
-                           analysis_file_key="state"
-                       ),
-                       CodingConfiguration(
-                           coding_mode=CodingModes.SINGLE,
-                           code_scheme=CodeSchemes.SOMALIA_ZONE,
-                           fold_strategy=FoldStrategies.assert_label_ids_equal,
-                           coded_field="zone_coded",
-                           analysis_file_key="zone",
-                       )
-                   ],
-                   code_imputation_function=code_imputation_functions.impute_somalia_location_codes,
-                   ws_code=CodeSchemes.WS_CORRECT_DATASET.get_code_with_match_value("location"),
-                   raw_field_fold_strategy=FoldStrategies.assert_equal)
-    ]
+    RQA_CODING_PLANS = []
+    DEMOG_CODING_PLANS = []
+    FOLLOW_UP_CODING_PLANS = []
+    SURVEY_CODING_PLANS = []
+    WS_CORRECT_DATASET_SCHEME = None
 
     def __init__(self, pipeline_name, raw_data_sources, phone_number_uuid_table, timestamp_remappings,
                  rapid_pro_key_remappings, project_start_date, project_end_date, filter_test_messages, move_ws_messages,
@@ -250,6 +59,13 @@ class PipelineConfiguration(object):
         self.drive_upload = drive_upload
         self.memory_profile_upload_url_prefix = memory_profile_upload_url_prefix
         self.data_archive_upload_url_prefix = data_archive_upload_url_prefix
+
+        PipelineConfiguration.RQA_CODING_PLANS = coding_plans.get_rqa_coding_plans(self.pipeline_name)
+        PipelineConfiguration.DEMOG_CODING_PLANS = coding_plans.get_demog_coding_plans(self.pipeline_name)
+        PipelineConfiguration.FOLLOW_UP_CODING_PLANS = coding_plans.get_follow_up_coding_plans(self.pipeline_name)
+        PipelineConfiguration.SURVEY_CODING_PLANS = PipelineConfiguration.DEMOG_CODING_PLANS + \
+                                                    PipelineConfiguration.FOLLOW_UP_CODING_PLANS
+        PipelineConfiguration.WS_CORRECT_DATASET_SCHEME = coding_plans.get_ws_correct_dataset_scheme(self.pipeline_name)
 
         self.validate()
 
@@ -573,7 +389,7 @@ class RapidProKeyRemapping(object):
 
 class DriveUpload(object):
     def __init__(self, drive_credentials_file_url, production_upload_path, messages_upload_path,
-                 individuals_upload_path, analysis_graphs_dir):
+                 individuals_upload_path, automated_analysis_dir):
         """
         :param drive_credentials_file_url: GS URL to the private credentials file for the Drive service account to use
                                            to upload the output files.
@@ -587,15 +403,15 @@ class DriveUpload(object):
         :param individuals_upload_path: Path in the Drive service account's "Shared with Me" directory to upload the
                                         individuals analysis CSV to.
         :type individuals_upload_path: str
-        :param analysis_graphs_dir: Directory in the Drive service account's "Shared with Me" directory to upload the
-                                    analysis graphs from this pipeline run to.
-        :type analysis_graphs_dir: str
+        :param automated_analysis_dir: Directory in the Drive service account's "Shared with Me" directory to upload the
+                                    automated analysis files from this pipeline run to.
+        :type automated_analysis_dir: str
         """
         self.drive_credentials_file_url = drive_credentials_file_url
         self.production_upload_path = production_upload_path
         self.messages_upload_path = messages_upload_path
         self.individuals_upload_path = individuals_upload_path
-        self.analysis_graphs_dir = analysis_graphs_dir
+        self.automated_analysis_dir = automated_analysis_dir
 
         self.validate()
 
@@ -605,10 +421,10 @@ class DriveUpload(object):
         production_upload_path = configuration_dict["ProductionUploadPath"]
         messages_upload_path = configuration_dict["MessagesUploadPath"]
         individuals_upload_path = configuration_dict["IndividualsUploadPath"]
-        analysis_graphs_dir = configuration_dict["AnalysisGraphsDir"]
+        automated_analysis_dir = configuration_dict["AutomatedAnalysisDir"]
 
         return cls(drive_credentials_file_url, production_upload_path, messages_upload_path,
-                   individuals_upload_path, analysis_graphs_dir)
+                   individuals_upload_path, automated_analysis_dir)
 
     def validate(self):
         validators.validate_string(self.drive_credentials_file_url, "drive_credentials_file_url")
@@ -618,4 +434,4 @@ class DriveUpload(object):
         validators.validate_string(self.production_upload_path, "production_upload_path")
         validators.validate_string(self.messages_upload_path, "messages_upload_path")
         validators.validate_string(self.individuals_upload_path, "individuals_upload_path")
-        validators.validate_string(self.analysis_graphs_dir, "analysis_graphs_dir")
+        validators.validate_string(self.automated_analysis_dir, "automated_analysis_dir")
